+++ src/routes/Redirect/Redirect.jsx
@@ -1,4 +1,4 @@
-import useLogin from '../hooks/useLogin';
+import useLogin from './hooks/useLogin';
 
 function Redirect() {
   useLogin();
